Add means of deterministic key pair generation
This is a somewhat naive implementation as it doesn't take into
consideration the differences between the "secp256k1" algorithm - used
to generate a regular key pair, and the "sha-256" algorithm - used to
generate the seeded key pair.

There is a theoretic concern that the "sha-256" algorithm doesn't have
the same spread in the full key space as the "secp256k1" has. The latter
is designed to avoid hash collisions as far as possible, the former is
designed to produce deterministic hashes from provided seeds.

Furthermore, the way the seeded key pair is intended to be used, the key
pairs it will produce are even more focused to a certain key sub-space
(the one targeting the human readable seed spectra). This may cause an
elevated risk of hash collisions.

We should investigate how big of an issue this is, if at all, and
address it accordingly.